added eventhandler bugtest (lazy user events), still fail!

+++ src/Array.spec.js
@@ -10,10 +10,11 @@ function compareEvent (is, should) {
   }
 }
 
-var numberOfYArrayTests = 30
-var repeatArrayTests = 1
+var numberOfYArrayTests = 10
+var repeatArrayTests = 20
 
 for (let database of databases) {
+  if (database !== 'memory') continue // TODO: REMOVE
   describe(`Array Type (DB: ${database})`, function () {
     var y1, y2, y3, yconfig1, yconfig2, yconfig3, flushAll
 
@@ -538,7 +539,17 @@ for (let database of databases) {
       }))
     })
     describeManyTimes(repeatArrayTests, `Random tests`, function () {
+      function fixAwaitingInType (array) {
+        array.eventHandler._tryCallEvents(array.eventHandler.awaiting)
+      }
       var randomArrayTransactions = [
+        function concurrentUserInteraction (array) {
+          if (array.eventHandler.awaiting === 0) {
+            array.eventHandler.awaiting = 1
+          } else {
+            fixAwaitingInType(array)
+          }
+        },
         function insert (array) {
           var c = getRandomNumber()
           var content = []
@@ -621,6 +632,7 @@ for (let database of databases) {
       it(`succeed after ${numberOfYArrayTests} actions, no GC, no disconnect`, async(function * (done) {
         yield applyRandomTransactionsNoGCNoDisconnect(this.users, this.arrays, randomArrayTransactions, numberOfYArrayTests)
         yield flushAll()
+        this.arrays.forEach(fixAwaitingInType)
         yield compareArrayValues(this.arrays)
         yield compareAllUsers(this.users)
         done()
@@ -628,6 +640,7 @@ for (let database of databases) {
       it(`succeed after ${numberOfYArrayTests} actions, no GC, all users disconnecting/reconnecting`, async(function * (done) {
         yield applyRandomTransactionsAllRejoinNoGC(this.users, this.arrays, randomArrayTransactions, numberOfYArrayTests)
         yield flushAll()
+        this.arrays.forEach(fixAwaitingInType)
         yield compareArrayValues(this.arrays)
         yield compareAllUsers(this.users)
         done()
@@ -635,6 +648,7 @@ for (let database of databases) {
       it(`succeed after ${numberOfYArrayTests} actions, GC, user[0] is not disconnecting`, async(function * (done) {
         yield applyRandomTransactionsWithGC(this.users, this.arrays, randomArrayTransactions, numberOfYArrayTests)
         yield flushAll()
+        this.arrays.forEach(fixAwaitingInType)
         yield compareArrayValues(this.arrays)
         yield compareAllUsers(this.users)
         done()
